Move `SchemaCreation` to its own file
